updated benchmarks

+++ include/hashing_project/helpers/probe_counts.cuh
@@ -3,7 +3,7 @@
 
 
 #if COUNT_PROBES
-__device__ __managed__ uint32_t helper_global_probes_count = 0;
+__device__ __managed__ uint64_t helper_global_probes_count = 0;
 #define ADD_PROBE_TILE \
   if (my_tile.thread_rank() == 0)   \
     atomicAdd(&helper_global_probes_count, 1);
@@ -15,11 +15,11 @@ __device__ __managed__ uint32_t helper_global_probes_count = 0;
 
 #define ADD_PROBE_BUCKET \
   if (my_tile.thread_rank() == 0)   \
-    atomicAdd(&helper_global_probes_count, (uint32_t) bucket_size*(sizeof(Key)+sizeof(Val))/128);
+    atomicAdd(&helper_global_probes_count, (uint64_t) bucket_size*(sizeof(Key)+sizeof(Val))/128);
 
 #define ADD_PROBE atomicAdd(&helper_global_probes_count, 1);
 namespace helpers {
-inline uint32_t get_num_probes() {
+inline uint64_t get_num_probes() {
   cudaDeviceSynchronize();
   auto count = helper_global_probes_count;
   helper_global_probes_count = 0;
@@ -33,7 +33,7 @@ inline uint32_t get_num_probes() {
 #define ADD_PROBE_BUCKET
 #define ADD_PROBE_ADJUSTED
 namespace helpers {
-inline uint32_t get_num_probes() {
+inline uint64_t get_num_probes() {
   return 0;
 }
 }
